Finish logger function, but leaving a tiny problem
Although the function of sensors reading and data logging is kind of
finished, there's a problem popping out that if I create a task with
RTOS on sensors_task, it crushes and panic without sign. However, if
I create the task with timer, it just work properly, so I don't get
what's the problem of the RTOS task issue.